Support timeout for "full" sensor execution time
Why? If the Airflow scheduler is restarted (e.g. due to
rollout of a new Kubernetes version), then the timeout
behaviour of a sensor should not be affected.
But before the code change, the timer would start from
zero again when the sensor is retried.

Solution: When a sensor is retried, then the sensor
uses the start date of the earliest try to justify
a time-out. To stay backwards-compatible, the new behaviour
is only active when `timeout_fullspan` is set to True.

Note: The exponential backoff feature for poking still uses
the start date of the current try. This is to keep the
code change small. No issues expected from that.